Fix linebreaks in vundlelog
Line-breaks internal to a log entry are replaced with null character when
writing to file, resulting in a single line log entry with a visual
representation of a null where new lines should start.

This changeset splits multiline log entries before writing, so that each
line appears as a different log entry. This makes the VundleLog more
readable.

Closes #7